fix: pwa build and test
