+++ .github/workflows/deploy.yml
@@ -3,7 +3,7 @@ name: Deploy to GitHub Pages
 on:
   push:
     branches:
-      - pwa-start
+      - pwa-build
 
 jobs:
   build-and-deploy:
