+++ index.html
@@ -9,23 +9,17 @@
     <!-- add favicon  -->
     <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
     <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
-    <link rel="icon" type="image/png" sizes="16x16" href="/icon-16x16.png" />
-    <link rel="icon" type="image/png" sizes="32x32" href="/icon-32x32.png" />
-    <link rel="icon" type="image/png" sizes="48x48" href="/icon-48x48.png" />
-    <link rel="icon" type="image/png" sizes="72x72" href="/icon-72x72.png" />
-    <link rel="icon" type="image/png" sizes="96x96" href="/icon-96x96.png" />
-    <link rel="icon" type="image/png" sizes="128x128" href="/icon-128x128.png" />
-    <link rel="icon" type="image/png" sizes="144x144" href="/icon-144x144.png" />
-    <link rel="icon" type="image/png" sizes="152x152" href="/icon-152x152.png" />
-    <link rel="icon" type="image/png" sizes="192x192" href="/icon-192x192.png" />
-    <link rel="icon" type="image/png" sizes="256x256" href="/icon-256x256.png" />
-    <link rel="icon" type="image/png" sizes="284x284" href="/icon-284x284.png" />
-    <link rel="icon" type="image/png" sizes="384x384" href="/icon-384x384.png" />
-    <link rel="icon" type="image/png" sizes="512x512" href="/icon-512x512.png" />
+    <link rel="icon" href="/icon-72x72.png" type="image/png" sizes="72x72" />
+    <link rel="icon" href="/icon-96x96.png" type="image/png" sizes="96x96" />
+    <link rel="icon" href="/icon-128x128.png" type="image/png" sizes="128x128" />
+    <link rel="icon" href="/icon-144x144.png" type="image/png" sizes="144x144" />
+    <link rel="icon" href="/icon-152x152.png" type="image/png" sizes="152x152" />
+    <link rel="icon" href="/icon-192x192.png" type="image/png" sizes="192x192" />
+    <link rel="icon" href="/icon-384x384.png" type="image/png" sizes="384x384" />
+    <link rel="icon" href="/icon-512x512.png" type="image/png" sizes="512x512" />
     <meta name="theme-color" content="%VITE_THEME_COLOR%" />
     <meta name="msapplication-TileImage" content="/icon-192x192.png" />
     <meta name="msapplication-TileColor" content="%VITE_THEME_COLOR%" />
-    <link rel="manifest" href="/manifest.json" />
     <!-- dynamic scripts -->
     <script type="module" src="/src/index.tsx"></script>
   </head>
