+++ src/assets/index.tsx
@@ -6,15 +6,6 @@ import globe from "./images/globe.png";
 import gluxy_cover from "./images/gluxy-cover.png";
 import heor_bg from "./images/herobg.png";
 import payment_card from "./images/payment-card.png";
-import planet01 from "./images/planet-01.png";
-import planet02 from "./images/planet-02.png";
-import planet03 from "./images/planet-03.png";
-import planet04 from "./images/planet-04.png";
-import planet05 from "./images/planet-05.png";
-import planet06 from "./images/planet-06.png";
-import planet07 from "./images/planet-07.png";
-import planet08 from "./images/planet-08.png";
-import planet09 from "./images/planet-09.png";
 import googel_play from "./images/play-google.svg";
 import roboto from "./images/robot.png";
 import send from "./images/send.svg";
@@ -36,16 +27,6 @@ export default {
   globe,
   gluxy_cover,
   heor_bg,
-  //planet
-  planet01,
-  planet02,
-  planet03,
-  planet04,
-  planet05,
-  planet06,
-  planet07,
-  planet08,
-  planet09,
   roboto,
   support,
   bg_hero,
