--- src/data/exploreing.tsx
@@ -1,29 +1,0 @@
-import assets from "../assets";
-
-export default [
-  {
-    id: "world-1",
-    imgUrl: assets.planet01,
-    title: "The Hogwarts",
-  },
-  {
-    id: "world-2",
-    imgUrl: assets.planet02,
-    title: "The Upside Down",
-  },
-  {
-    id: "world-3",
-    imgUrl: assets.planet03,
-    title: "Kadirojo Permai",
-  },
-  {
-    id: "world-4",
-    imgUrl: assets.planet04,
-    title: "Paradise Island",
-  },
-  {
-    id: "world-5",
-    imgUrl: assets.planet05,
-    title: "Hawkins Labs",
-  },
-];
