+++ src/data/logo.tsx
@@ -1,6 +1,5 @@
 export default {
   default: {
-    type: "image/png",
     src: "/icon-512x512.png",
   },
 };
