+++ tsconfig.node.json
@@ -6,5 +6,5 @@
     "moduleResolution": "bundler",
     "allowSyntheticDefaultImports": true
   },
-  "include": ["vite.config.ts"]
+  "include": ["vite-config/*.ts", "vite-config/*.ts", "vite/*.ts", "vite.config.ts"]
 }
