+++ vite-config/global.ts
@@ -1,0 +1,5 @@
+import { loadEnv } from "vite";
+
+try {
+  Object.assign(process.env, loadEnv("mock", process.cwd(), ""));
+} catch (error) {}
